fix(model): move `Member.kt` from ``model dir`` to ``entity dir`` - remove model dir
