+++ src/main/kotlin/techcourse/herobeans/entity/Member.kt
@@ -1,4 +1,4 @@
-package techcourse.herobeans.model
+package techcourse.herobeans.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
