+++ src/main/kotlin/techcourse/herobeans/repository/MemberJpaRepository.kt
@@ -1,7 +1,7 @@
 package techcourse.herobeans.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
-import techcourse.herobeans.model.Member
+import techcourse.herobeans.entity.Member
 
 interface MemberJpaRepository : JpaRepository<Member, Long> {
     fun findByEmail(email: String): Member?
